move to legacy: eager
