+++ weave/legacy/eager.py
@@ -2,8 +2,8 @@ import typing
 
 from weave.legacy import context_state, graph
 
-from . import weave_internal
-from . import weave_types as types
+from .. import weave_internal
+from .. import weave_types as types
 
 WeaveIterObjectType = typing.TypeVar("WeaveIterObjectType")
